CORE-6645 Fixed data-info's directory create endpoint.
Using (juxt filter remove) to split existing paths from paths to create did not allow the (doseq … set-owner …) call to properly evaluate the list of paths that were created, so ownership was not set for the given user on the created paths.
Updated this logic to use group-by, which does evaluate the entire path list into vectors.
Removed debug logging statements.