CORE-7620 CORE-7621 Refactor data-info to use metadata-client.
The data-info service AVU endpoints will now use the new metadata-client
to get and set metadata AVUs in addition to iRODS AVUs.
So the logic of coordinating data-info and metadata endpoints will be
migrated out of terrain in following commit(s), and some of it has
already been migrated into data-info with this commit.
Refactored data-info set/add AVU endpoints to POST/PUT endpoints, so
they're consistent with the corresponding metadata endpoints.