Create optimized uuid-for-path endpoint in data-info
This endpoint uses more direct methods for getting the UUID, for cases
where a full stat entry isn't needed, just the path. Terrain has also
been updated to use this for the data-info client's uuid-for-path
function, which is used for a lot of terrain's current translating of
paths into uuids.

Ideally, this eventually goes away or nearly away as we move the UI to
use data UUIDs in more/most places, such that terrain doesn't need to be
making these lookups. But, for now, this should trim off a decent amount
off each request it has to make to resolve a UUID.